# backend/openapi/applyflow-v1.yaml
openapi: 3.0.3
info:
  title: ApplyFlow API
  version: v1
  description: The versioned API for the ApplyFlow portfolio application.
paths:
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [Authentication]
      summary: Create a private user account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '200':
          description: Account created; an HttpOnly refresh cookie is also established
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Authentication]
      summary: Sign in with an email address and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated; an HttpOnly refresh cookie is also established
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      tags: [Authentication]
      summary: Restore an access token from the refresh cookie
      responses:
        '200':
          description: Identity restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [Authentication]
      summary: Revoke the current refresh session and clear its cookie
      responses:
        '204': { description: Signed out }
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Authentication]
      summary: Retrieve the user represented by the access token
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current private user context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/system/status:
    get:
      operationId: getSystemStatus
      tags: [System]
      summary: Check the API and database path
      responses:
        '200':
          description: The application and migrated database are available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        default:
          description: An RFC 9457 API problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: An RFC 9457 API problem
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, maxLength: 128 }
    CurrentUser:
      type: object
      required: [id, email, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [accessToken, expiresIn, user]
      properties:
        accessToken: { type: string, description: A 15-minute JWT held by the SPA in memory }
        expiresIn: { type: integer, format: int64, example: 900 }
        user: { $ref: '#/components/schemas/CurrentUser' }
    SystemStatus:
      type: object
      required: [status, version, database]
      properties:
        status:
          type: string
          example: ok
        version:
          type: string
          example: 0.1.0
        database:
          type: string
          example: connected
    ProblemDetail:
      type: object
      required: [type, title, status, detail, instance, code, timestamp, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        timestamp: { type: string, format: date-time }
        traceId: { type: string, format: uuid }
        fieldErrors:
          type: object
          additionalProperties: { type: string }